# perf(serialize): use quick sort for sorting everything