We can use `quickSort` for sorting `Set` and other arrays too.

### Compared to target: https://github.com/unjs/ohash/pull/148

```sc
 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:small 2259ms
     name                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ 8434bea  3,842,100.95  0.0002  3.0409  0.0003  0.0003  0.0005  0.0006  0.0007  ±1.21%  1921051
   · ohash @ dev      3,878,229.53  0.0002  2.0584  0.0003  0.0003  0.0004  0.0005  0.0008  ±0.86%  1939115   fastest

 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:large 1219ms
     name                    hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ 8434bea  13,574.99  0.0626  2.9758  0.0737  0.0736  0.1290  0.1816  0.2739  ±1.27%     6788
   · ohash @ dev      46,216.51  0.0190  4.3283  0.0216  0.0208  0.0426  0.0517  0.1531  ±1.76%    23109   fastest

 BENCH  Summary

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:small
    1.01x faster than ohash @ 8434bea

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:large
    3.40x faster than ohash @ 8434bea
```

```sc
clk: ~5.13 GHz
cpu: AMD Ryzen 9 7950X 16-Core Processor
runtime: bun 1.2.4 (x64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
• count:1, size:small
------------------------------------------- -------------------------------
ohash @ 8434bea              188.35 ns/iter 180.67 ns ▇█
                    (164.72 ns … 506.32 ns) 466.47 ns ██
                    (  0.00  b … 360.00  b)   4.18  b ██▄▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

ohash @ dev                  204.06 ns/iter 191.49 ns  █
                    (176.71 ns … 547.58 ns) 449.94 ns ▂█
                    (  0.00  b … 924.00  b)  13.36  b ██▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

summary
  ohash @ 8434bea
   1.08x faster than ohash @ dev

• count:1, size:large
------------------------------------------- -------------------------------
ohash @ 8434bea               73.83 µs/iter  76.12 µs   █
                       (61.11 µs … 1.48 ms) 126.72 µs  ██
                    (  0.00  b … 528.00 kb)   9.96 kb ▂███▅▅▄▃▃▂▂▂▂▂▁▁▁▁▁▁▁

ohash @ dev                   28.10 µs/iter  29.14 µs  █
                       (21.46 µs … 1.46 ms)  60.65 µs  █
                    (  0.00  b … 264.00 kb)   2.82 kb ▂██▂▄▃▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁

summary
  ohash @ dev
   2.63x faster than ohash @ 8434bea