@zsilbi 2025-03-08
### Compared the previous versions:

```sc
 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:small 2839ms
     name                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ v1.1.6     973,475.82  0.0008  0.3759  0.0010  0.0010  0.0017  0.0023  0.0093  ±0.67%   486738   slowest
   · ohash @ v2.0.11  2,381,447.00  0.0003  3.3053  0.0004  0.0004  0.0007  0.0008  0.0099  ±1.33%  1190724
   · ohash @ dev      3,910,502.92  0.0002  0.4540  0.0003  0.0003  0.0004  0.0005  0.0007  ±0.32%  1955252   fastest

 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:large 1820ms
     name                    hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ v1.1.6    7,775.26  0.1052  2.6945  0.1286  0.1229  0.3188  0.3486  0.5284  ±1.36%     3888   slowest
   · ohash @ v2.0.11  10,418.69  0.0886  0.3633  0.0960  0.0948  0.1351  0.2645  0.3105  ±0.49%     5210
   · ohash @ dev      46,801.00  0.0189  0.3595  0.0214  0.0208  0.0383  0.0491  0.1652  ±0.51%    23401   fastest

 BENCH  Summary

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:small
    1.64x faster than ohash @ v2.0.11
    4.02x faster than ohash @ v1.1.6

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:large
    4.49x faster than ohash @ v2.0.11
    6.02x faster than ohash @ v1.1.6
```

```sc
clk: ~5.29 GHz
cpu: AMD Ryzen 9 7950X 16-Core Processor
runtime: bun 1.2.4 (x64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
• count:1, size:small
------------------------------------------- -------------------------------
ohash @ v1.1.6               899.12 ns/iter 897.80 ns  █
                      (835.54 ns … 1.24 µs)   1.17 µs ▇█▆
                    (  0.00  b … 990.00  b)  29.82  b ███▅▃▂▂▃▃▃▁▂▃▂▂▂▂▂▂▁▂

ohash @ v2.0.11              480.68 ns/iter 461.10 ns  █
                    (443.97 ns … 882.91 ns) 795.39 ns ██
                    (  0.00  b …   1.48 kb)  46.84  b ██▂▂▁▁▁▂▁▁▁▂▁▂▁▁▁▁▁▁▁

ohash @ dev                  196.69 ns/iter 185.76 ns  █
                    (173.12 ns … 572.94 ns) 454.37 ns ▇█
                    (  0.00  b … 726.00  b)  11.86  b ██▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

summary
  ohash @ dev
   2.44x faster than ohash @ v2.0.11
   4.57x faster than ohash @ v1.1.6

• count:1, size:large
------------------------------------------- -------------------------------
ohash @ v1.1.6               228.55 µs/iter 220.34 µs  █
                      (182.91 µs … 1.79 ms) 804.14 µs ▅█
                    (  0.00  b …   1.03 mb)  17.08 kb ██▆▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

ohash @ v2.0.11              108.31 µs/iter 111.17 µs    █
                       (88.85 µs … 1.47 ms) 171.21 µs   ██▄
                    (  0.00  b … 528.00 kb)  19.06 kb ▁▆███▅▅▅▄▃▃▂▂▁▁▁▁▁▁▁▁

ohash @ dev                   27.06 µs/iter  27.76 µs  ▆█
                       (20.83 µs … 1.48 ms)  51.95 µs  ██
                    (  0.00  b … 264.00 kb)   2.92 kb ▂██▆▃▃▆▃▂▂▂▂▂▁▁▁▁▁▁▁▁

summary
  ohash @ dev
   4x faster than ohash @ v2.0.11
   8.45x faster than ohash @ v1.1.6
```